dev: check input image types and relocate struct definition
